[REFACTOR] Fixed tests [LUNN/BARREYRE]
