+++ app/src/androidTest/java/ca/dal/cs/athletemonitor/athletemonitor/LoginActivityTest.java
@@ -21,69 +21,13 @@ import static android.support.test.espresso.matcher.ViewMatchers.withText;
 
 //import for intended() and hasComponent()
 import static android.support.test.espresso.intent.Intents.intended;
-import static android.support.test.espresso.intent.matcher.IntentMatchers.ha
-
-
 
 /**
  * UI Test for Login Activity
  */
-
 public class LoginActivityTest {
-    private String usernameInputTestText = "username";
-    private String passwordInputTestText = "password";
-
-
-    @Rule
-    public ActivityTestRule<LoginActivity> mActivityRule = new ActivityTestRule<LoginActivity>(
-            LoginActivity.class);
     @Rule
     public IntentsTestRule<LoginActivity> intentsTestRule = new IntentsTestRule<>(LoginActivity.class);
-
-
-    /**
-     * Test username with empty string
-     *
-     * @throws Exception
-     */
-    @Test
-    public void usernameEmptyTest() throws Exception {
-        onView(withId(R.id.txtUsername)).check(matches(withText("")));
-    }
-
-    /**
-     * Test username textbox with input
-     *
-     * @throws Exception
-     */
-    @Test
-    public void usernameInputTextTest() throws Exception {
-        onView(withId(R.id.txtUsername)).perform(typeText(usernameInputTestText));
-        closeSoftKeyboard();
-        onView(withId(R.id.txtUsername)).check(matches(withText(usernameInputTestText)));
-    }
-
-    /**
-     * Test password with empty string
-     *
-     * @throws Exception
-     */
-    @Test
-    public void passwordEmptyTest() throws Exception {
-        onView(withId(R.id.txtPassword)).check(matches(withText("")));
-    }
-
-    /**
-     * Test username textbox with input
-     *
-     * @throws Exception
-     */
-    @Test
-    public void passwordInputTextTest() throws Exception {
-        onView(withId(R.id.txtPassword)).perform(typeText(passwordInputTestText));
-        closeSoftKeyboard();
-        onView(withId(R.id.txtPassword)).check(matches(withText(passwordInputTestText)));
-    }
 
     /**
      * Test successful login when signin button is clicked
@@ -93,26 +37,9 @@ public class LoginActivityTest {
     @Test
     public void signInSuccessTest() throws Exception {
         //perform signin button click
-        onView(withId(R.id.btnSignin)).perform(click());
+        onView(withId(R.id.btnSignIn)).perform(click());
 
         //test if main activity has been launched
         intended(hasComponent(MainActivity.class.getName()));
-
-        //test correct response on successful login
-        //onView(withId(R.id.lblLoginResult)).check(matches(withText(R.string.loginSuccess)));
-        //onView(withId(R.id.btnSignin)).check(matches(withText(R.string.loginSuccess)));
-
-
     }
-
-
-
-
 }
-
-//
-//    // invalid password
-//    onView(withId(R.id.txtPassword)).perform(typeText(oneRulePassedInput));
-//    closeSoftKeyboard();
-//    onView(withId(R.id.btnValidate)).perform(click());
-//    onView(withId(R.id.txtviewPasswordStrength)).check(matches(withText(oneRulePassedResponse)));
